form for entrys finished

+++ src/server/api/routers/punishmentsAndContributionType.ts
@@ -7,6 +7,7 @@ import {
   protectedProcedure,
   publicProcedure,
 } from "~/server/api/trpc";
+import { AuthUtil } from "~/utils/auth-utils";
 
 export const punishmentOrContributionTypeRouter = createTRPCRouter({
   create: protectedProcedure
@@ -14,11 +15,18 @@ export const punishmentOrContributionTypeRouter = createTRPCRouter({
       z.object({
         name: z.string(),
         description: z.string(),
-        price: z.number(),
+        price: z.number().optional(),
         teamId: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
+      const isUserManager = await AuthUtil.isUserManager(
+        ctx.prisma,
+        input.teamId,
+        ctx.auth.userId
+      );
+
+      if (!isUserManager) throw new TRPCError({ code: "FORBIDDEN" });
       const data = await ctx.prisma.punishmentOrContributionType.create({
         data: {
           name: input.name,
@@ -34,9 +42,17 @@ export const punishmentOrContributionTypeRouter = createTRPCRouter({
     .input(
       z.object({
         punishmentOrContributionId: z.string(),
+        teamId: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
+      const isUserManager = await AuthUtil.isUserManager(
+        ctx.prisma,
+        input.teamId,
+        ctx.auth.userId
+      );
+
+      if (!isUserManager) throw new TRPCError({ code: "FORBIDDEN" });
       return await ctx.prisma.punishmentOrContributionType.delete({
         where: {
           id: input.punishmentOrContributionId,
@@ -51,9 +67,18 @@ export const punishmentOrContributionTypeRouter = createTRPCRouter({
         description: z.string().optional(),
         price: z.number().optional(),
         punishmentOrContributionId: z.string(),
+        teamId: z.string(),
       })
     )
     .mutation(async ({ ctx, input }) => {
+      const isUserManager = await AuthUtil.isUserManager(
+        ctx.prisma,
+        input.teamId,
+        ctx.auth.userId
+      );
+
+      if (!isUserManager) throw new TRPCError({ code: "FORBIDDEN" });
+
       return await ctx.prisma.punishmentOrContributionType.update({
         data: {
           name: input.name,
@@ -73,14 +98,22 @@ export const punishmentOrContributionTypeRouter = createTRPCRouter({
       })
     )
     .query(async ({ ctx, input }) => {
+      const isUserInTeam = await AuthUtil.isUserTeamMember(
+        ctx.prisma,
+        input.teamId,
+        ctx.auth.userId
+      );
+      if (!isUserInTeam) throw new TRPCError({ code: "FORBIDDEN" });
       const list = await ctx.prisma.punishmentOrContributionType.findMany({
         where: {
           teamId: input.teamId,
         },
       });
-      const isUserManager = await ctx.prisma.teamManager.findFirst({
-        where: { teamId: input.teamId, clerkId: ctx.auth.userId },
-      });
+      const isUserManager = await AuthUtil.isUserManager(
+        ctx.prisma,
+        input.teamId,
+        ctx.auth.userId
+      );
 
       return {
         list,
